Refactored code comparer class, removed code generator. Adjusted spacing in game file
